- Add gmp utilities for integer time handling in Events - Update python => ok
git-svn-id: svn+ssh://gforge/svn/siconos/trunk@721 bf01a9d7-0df3-0310-8ad4-b1c94d9ae2d5

+++ Kernel/src/controlTools/ActuatorEvent.h
@@ -48,10 +48,10 @@ private:
 public:
 
   /** constructor with time value as a parameter
-   *  \param an unsigned long int
+   *  \param a double
    *  \param a string, type of Event
    */
-  ActuatorEvent(unsigned long int, const std::string&);
+  ActuatorEvent(double, const std::string&);
 
   /** destructor
    */
